requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==2.0.0
Flask==2.0.1
Flask-WTF==0.15.1
gunicorn==20.1.0
idna==3.2
ipaddr==2.2.0
ipwhois==1.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
requests==2.19.0
simplejson==3.17.2
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
